Add meetup ICS
This is valid per https://icalendar.org/validator.html and works in my
Apple Calendar. The only issue is that the first recurring event
overlaps with the last scheduled event. To avoid this, we probably need
to generate the ICS file dynamically in Ruby. I'm outta time for now,
but PRs are welcome!